Introduce manifest_archives.json and update service worker caching infrastructure.
This is the first step of manifest splitting, which extracts non-essential
content from the main manifest out into separate files that would be loaded
on demand. The main manifest will eventually contain only the information
required for the first page load and include versioning information of other
extracted manifests, so that those separate manifests will not reload until
they are modified, which mitigates potential performance degrades caused by
this splitting.

Service worker code is overhauled and caching behavior is refined meanwhile.

The consumer of archives in js/buses.js will be modified to use the new separate
manifest_archives.json in a future commit.